src/backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
httpx
python-dotenv
